+++ src/Mapper/EditorInterface.php
@@ -28,7 +28,7 @@ class EditorInterface extends BaseMapper
      */
     public function map($resource, array $data): ResourceClass
     {
-        return $this->hydrate($resource ?: ResourceClass::class, [
+        return $this->hydrator->hydrate($resource ?: ResourceClass::class, [
             'sys' => new SystemProperties($data['sys']),
             'controls' => \array_map([$this, 'mapControl'], $data['controls']),
         ]);
@@ -41,7 +41,7 @@ class EditorInterface extends BaseMapper
      */
     protected function mapControl(array $data): Control
     {
-        return $this->hydrate(Control::class, [
+        return $this->hydrator->hydrate(Control::class, [
             'fieldId' => $data['fieldId'],
             'widgetId' => $data['widgetId'],
             'settings' => $data['settings'] ?? [],