+++ src/Mapper/Extension.php
@@ -28,7 +28,7 @@ class Extension extends BaseMapper
      */
     public function map($resource, array $data): ResourceClass
     {
-        return $this->hydrate($resource ?: ResourceClass::class, [
+        return $this->hydrator->hydrate($resource ?: ResourceClass::class, [
             'sys' => new SystemProperties($data['sys']),
             'name' => $data['extension']['name'],
             'source' => $data['extension']['src'] ?? $data['extension']['srcdoc'] ?? '',