+++ src/Mapper/PersonalAccessToken.php
@@ -28,7 +28,7 @@ class PersonalAccessToken extends BaseMapper
      */
     public function map($resource, array $data): ResourceClass
     {
-        return $this->hydrate($resource ?: ResourceClass::class, [
+        return $this->hydrator->hydrate($resource ?: ResourceClass::class, [
             'sys' => new SystemProperties($data['sys']),
             'name' => $data['name'],
             'isReadOnly' => !\in_array('content_management_manage', $data['scopes'], \true),