+++ src/Mapper/PreviewApiKey.php
@@ -35,7 +35,7 @@ class PreviewApiKey extends BaseMapper
             ));
         }
 
-        return $this->hydrate(ResourceClass::class, [
+        return $this->hydrator->hydrate(ResourceClass::class, [
             'sys' => new SystemProperties($data['sys']),
             'name' => $data['name'],
             'accessToken' => $data['accessToken'],