+++ src/Mapper/WebhookHealth.php
@@ -34,7 +34,7 @@ class WebhookHealth extends BaseMapper
             ));
         }
 
-        return $this->hydrate(ResourceClass::class, [
+        return $this->hydrator->hydrate(ResourceClass::class, [
             'sys' => new SystemProperties($data['sys']),
             'total' => $data['calls']['total'],
             'healthy' => $data['calls']['healthy'],